pyproject.toml
[project]
name = "loan-underwriting-env"
version = "1.0.0"
description = "RL environment for AI loan underwriting agents"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai>=1.0.0",
    "python-dotenv"
]

[project.scripts]
server = "main:app"

[project.urls]
Homepage = "https://huggingface.co/spaces/subhasreeee/loan-underwriting-env"

[tool.openenv]
name = "loan-underwriting-env"
version = "1.0.0"
benchmark = "loan-underwriting"
tasks = ["task_easy", "task_medium", "task_hard", "task_batch"]